fix: 修复图片生成size not support问题
Fixes #1224, #1068
